backend/requirements.txt
asgiref==3.8.1
Django==5.2.3
django-rest-framework==0.1.0
djangorestframework==3.16.0
pillow==11.2.1
sqlparse==0.5.3
